Apply suggestions from code review

+++ gamemodes/terrortown/entities/entities/randomat_hellosanta_coal/shared.lua
@@ -23,6 +23,8 @@ if SERVER then
         local owner = self:GetOwner()
 
         if not IsValid(ent) or not ent:IsPlayer() or not ent:IsActive() then return end
+        -- Don't let this kill someone if it's just being carried by a magneto stick
+        if IsValid(phys) and phys:HasGameFlag(FVPHYSICS_PLAYER_HELD) then return end
         if data.Speed < 300 then return end -- The coal has to be going fast enough to kill someone
 
         if CurTime() <= self.antiSpam then return end
@@ -33,7 +35,7 @@ if SERVER then
         self.lifetime = CurTime() + 3 -- Leave the coal around for a few more seconds then remove it
         if Randomat:IsTraitorTeam(ent) or (Randomat:IsJesterTeam(ent) and GetConVar("randomat_hellosanta_jesters_are_naughty"):GetBool()) or (Randomat:IsIndependentTeam(ent) and GetConVar("randomat_hellosanta_independents_are_naughty"):GetBool()) then
             owner:PrintMessage(HUD_PRINTTALK, ent:Nick() .. " was naughty and your ammo has been refunded.")
-            owner:SetNWBool("XmasCannonHasAmmo", true)
+            owner:SetNWBool("RdmtXmasCannonHasAmmo", true)
         else
             owner:PrintMessage(HUD_PRINTTALK, ent:Nick() .. " was nice and the christmas cannon has been disabled.")
         end
